requirements.txt
Django>=3.2.0,<3.3.0
djangorestframework>=3.13.0,<3.14.0
psycopg2>=2.7.5,<2.8.0
django-extensions>=3.1.0,<3.2.0

flake8>=3.6.0,<3.7.0
